requirements.txt
fastapi>=0.104.0
pydantic>=2.0.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
crewai[tools]>=0.86.0,<1.0.0 